Added a "NoFlash" Playback
This will detect if the browser requires flash to playback the required
media when there is no HTML5/JS support. If flash isn't installed on the
browser, then a message pertaining to adobe flash will be displayed
instead of "Your browser does not support the playback of this video.
Try to use a different browser."!