Refactor BUFFERING and BUFFERFULL events and added 'buffering'
PLAYBACK_BUFFERING and PLAYBACK_BUFFERFULL are now only fired whenever
the new 'buffering' state changes.

The buffering event is fired when loading has started the tech isn't
happy that it's got enough content to play smoothly. The bufferfull
event is fired when the tech is happy it's got enough content to play
smoothly.

The readyState in chrome (didn't test firefox) remained at
"HAVE_ENOUGH_DATA" even when it had was stuttering/stalled, so there's
also a separate check to determine if the playhead is actually moving.
'buffering' is triggered if the playhead isn't moving but the tech is
stating that it should be playing.

PLAYBACK_READY is now fired when rendering finishes, as at this point
'play()' can actually be called successfully and it means it's the same
for html5_video as hls.