package: move all build dependencies to devDependencies
This will avoid issues for projects that include 
clappr as dependency but will use its compiled
version directly, instead of recompiling it and
bundling it in a single js file.

This is probably the most common use for projects
that do not use webpack as their build system.
Perhaps adding instructions to allow projects to
compile clappr by source would be useful.