Ensure dvr only enables when duration has stabilized
The duration will initially be the actual duration until the real
duration is longer the hlsjs's live sync point, then it drops back down
to hide the content after the sync point (to prevent someone seeking to
a point which will cause their buffer to run out before the next chunk
arrives).

This ensures that the dvr can only be enabled when the duration is not
including the content after the live sync point.