Remove old baseUrl references and provide way to set baseUrl at runtime
It looked like where baseUrl was referenced before it never had any
effect, as this changed when we switched to webpack.

The proper way of setting the assets base url at runtime is in the
webpack documentation:
https://webpack.github.io/docs/configuration.html#output-publicpath
and I have added an extra entry point which looks to see if the user
wants a custom base url, and sets it if they do before the main app
loads.
This uses a global because I don't think there's another way.