Html5Video: remove src on stop() and readd on play()
Removed adding <source> as I don't think this was necessary. I think the
browser would look at the source list to detemrine what url to assign to
the video tag itself, which we are always doing in setupUrl(). The video
tag itself doesn't have a 'type' attribute which is why I think this.