feat(mediator): drop mediator support
The Mediator is a helper once used for bridging Flash implementations to their Clappr component
counterparts. Since Flash is almost completely dropped for the majority of browsers and the Flash
components are being less and less used, it's better to remove it since it can create lots of
potential issues.

BREAKING CHANGE: Components that use the Mediator will no longer work on Clappr. Since this is a
major breaking change, a new version check will be added to the Loader to prevent issues with
existing components.