feat(playerinfo): drop playerinfo support
The PlayerInfo is a global object that holds information about all player instances, their
playbacks, sizes and options. It was used to provide some information throughout contexts, but it
makes no sense to have it now. Instead we'll use resize events, local scope options and if the
playback list is necessary, check the Loader instance.

BREAKING CHANGE: Components that use the PlayerInfo will no longer work on Clappr. Since this is
a major breaking change, the version check from the Loader should be used to restrict plugins
that support this change.