test(base_object): update existent test
BaseObject class uses the uniqueId module to generate unique ids. As the same instance of the uniqueId module is being used within the BaseObject class the reference of unique ids created remains between tests. Therefore, when another test is added, the unique ids of the existing tests are changed too.