refactor: destroy old shaka-player instance when setup occurs
One error (code 3015) occurs when the setInterval loop on _startTimeUpdateTimer method tries to update the current time of the old shaka-player instance in the time interval between the _setup method is called and one new shaka-player instance is created.

This code guarantees the delete of the old shaka-player instance before the creation of one new one into the _setup method.